requirements.txt
wheel
bob-telegram-tools==1.0.0
dataclasses==0.7
dcase-util==0.2.11
h5py==2.10.0
imbalanced-learn==0.6.2
keras==2.2.4
keras-efficientnets==0.1.7
librosa==0.7.2
matplotlib==3.1.3
nlpaug==1.0.1
numba==0.48.0
numpy==1.19.1
pandas==0.25.3
pyaml==20.4.0
scikit-learn==0.22.2.post1
scipy==1.5.2
sed-eval==0.2.1
soundfile==0.10.3.post1
#tensorboard==2.2.2
tensorboardX==2.1
tensorflow-gpu==1.13.1
tensorflow-probability==0.6.0
tqdm==4.43.0
#skopt==0.8.1
keras-multi-head==0.27.0
